pyproject.toml
[tool.poetry]
name = "hasura-dapr-background-job"
version = "0.1.0"
description = ""
authors = ["Nakano as a Service <nakanoasaservice@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.10"
dapr-ext-grpc = "^1.7.0"
cloudevents = "^1.5.0"
requests = "^2.28.1"
pydantic = "^1.9.2"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
